+++ Src/Public/New-AsBuiltConfig.ps1
@@ -1,17 +1,20 @@
 function New-AsBuiltConfig {
     <#
     .SYNOPSIS  
-        Creates As Built JSON configuration files.
+        Creates As Built Report configuration files.
     .DESCRIPTION
         New-AsBuiltConfig starts a menu-driven procedure in the powershell console and asks the user a series of questions
         Answers to these questions are optionally saved in a JSON configuration file which can then be referenced using the
-        -AsBuiltConfigPath parameter using New-AsBuiltReport, to save having to answer these questions again and also to allow
+        -AsBuiltConfigFilePath parameter using New-AsBuiltReport, to save having to answer these questions again and also to allow
         the automation of New-AsBuiltReport.
         
-        New-AsBuiltConfig will automatically be called by New-AsBuiltReport if the -AsBuiltConfigPath parameter is not specified
-        If a user wants to generate a new As Built JSON configuration without running a new report, this cmdlet is exported
+        New-AsBuiltConfig will automatically be called by New-AsBuiltReport if the -AsBuiltConfigFilePath parameter is not specified
+        If a user wants to generate a new As Built Report configuration without running a new report, this cmdlet is exported
         in the AsBuiltReport powershell module and can be called as a standalone cmdlet.
     #>
+
+    [CmdletBinding()]
+    param()
 
     #Run section to prompt user for information about the As Built Report to be exported to JSON format (if saved)
     $global:Config = @{ }
@@ -21,7 +24,7 @@ function New-AsBuiltConfig {
     Write-Host '---------------------------------------------' -ForegroundColor Cyan
     Write-Host ' <        As Built Report Information      > ' -ForegroundColor Cyan
     Write-Host '---------------------------------------------' -ForegroundColor Cyan
-    $ReportAuthor = Read-Host -Prompt "Enter the name of the Author for this As Built report [$env:USERNAME]"
+    $ReportAuthor = Read-Host -Prompt "Enter the name of the Author for this As Built Report [$env:USERNAME]"
     if (($ReportAuthor -like $null) -or ($ReportAuthor -eq "")) {
         $ReportAuthor = $env:USERNAME
     }
@@ -37,9 +40,9 @@ function New-AsBuiltConfig {
     Write-Host ' <           Company Information           > ' -ForegroundColor Cyan
     Write-Host '---------------------------------------------' -ForegroundColor Cyan
 
-    $CompanyInfo = Read-Host -Prompt "Would you like to enter Company information for the As Built report? (y/n)"
+    $CompanyInfo = Read-Host -Prompt "Would you like to enter Company information for the As Built Report? (y/n)"
     while ("y", "n" -notcontains $CompanyInfo) {
-        $CompanyInfo = Read-Host -Prompt "Would you like to enter Company information for the As Built report? (y/n)"
+        $CompanyInfo = Read-Host -Prompt "Would you like to enter Company information for the As Built Report? (y/n)"
     }
 
     if ($CompanyInfo -eq 'y') {
@@ -137,58 +140,52 @@ function New-AsBuiltConfig {
     #endregion Email Configuration
 
     #region Report Configuration Folder
-    if (!($ReportConfigPath)) {
+    if ($Report -and !$ReportConfigFilePath) {
         Clear-Host
         Write-Host '---------------------------------------------' -ForegroundColor Cyan
         Write-Host ' <          Report Configuration           > ' -ForegroundColor Cyan
         Write-Host '---------------------------------------------' -ForegroundColor Cyan
-        $ReportConfigFolder = Read-Host -Prompt "Enter the full path of the folder to use for storing report JSON configuration files and custom style scripts [$env:USERPROFILE\AsBuiltReport]"
+        $ReportConfigFolder = Read-Host -Prompt "Enter the full path of the folder to use for storing report configuration files and custom style scripts [$env:USERPROFILE\AsBuiltReport]"
         if (($ReportConfigFolder -like $null) -or ($ReportConfigFolder -eq "")) {
             $ReportConfigFolder = $env:USERPROFILE + "\AsBuiltReport"
         }
 
         #If the folder doesn't exist, create it
         if (!(Test-Path -Path $ReportConfigFolder)) {
-            New-Item -Path $ReportConfigFolder -ItemType Directory -Force
+            Try {
+                $Folder = New-Item -Path $ReportConfigFolder -ItemType Directory -Force
+            } Catch {
+                Write-Error $_
+                break
+            }
         }
 
-        #Add the path to the folder to the JSON configuration
+        #Add the path to the folder to the report configuration file
         $Config.UserFolder = @{
             'Path' = $ReportConfigFolder
         }
 
-        #Test to see if the Report Configuration folder is empty. If it is, generate all report JSON files.
-        #If the folder is not empty, do a foreach loop through each currently installed report module and check if the
-        #report json specific to that module exists. If it does, prompt the user to see if they want to overwrite the
-        #JSON. If it doesn't exist, generate the JSON
-
-        $AsBuiltReportModules = Get-Module -Name "AsBuiltReport.*" -ListAvailable | Where-Object { $_.name -ne 'AsBuiltReport.Core' } | Sort-Object -Property Version -Descending | Select-Object -Unique
-        if (!(Get-ChildItem -Path $ReportConfigFolder -Force)) {
-            Foreach ($AsBuiltReportModule in $AsBuiltReportModules) {
-                $AsBuiltReportName = $AsBuiltReportModule.Name.Replace("AsBuiltReport.", "")
-                Try {
-                    New-AsBuiltReportConfig -Report $AsBuiltReportName -Path $ReportConfigFolder
-                } Catch {
-                    Write-Error $_
-                    Break
-                }
-            }
+        # Test to see if the report configuration file exists. If it doesn't exist, generate the report configuration file.
+        # If the report configuration file exists, prompt the user to overwrite the report configuration file.
+        $ReportModule = Get-Module -Name "AsBuiltReport.$Report" -ListAvailable | Sort-Object -Property Version -Descending | Select-Object -First 1
+        $ReportModulePath = "$($ReportConfigFolder)\$($ReportModule.Name).json"
+        if (!(Get-ChildItem -Path $ReportModulePath)) {
+            Write-Verbose -Message "Copying '$($ReportModule.ModuleBase)\$($ReportModule.Name).json' to '$($ReportModulePath)'."
+            New-AsBuiltReportConfig -Report $Report -FolderPath $ReportConfigFolder
         } else {
             try {
-                foreach ($AsBuiltReportModule in $AsBuiltReportModules) {
-                    $AsBuiltReportName = $AsBuiltReportModule.Name.Replace("AsBuiltReport.", "")
-                    if (Test-Path -Path "$($ReportConfigFolder)\$($AsBuiltReportModule.Name).json") {
-                        $OverwriteReportJSON = Read-Host -Prompt "A report JSON already exists in the specified folder for $($AsBuiltReportModule.Name). Would you like to overwrite it? (y/n)"
-                        while ("y", "n" -notcontains $OverwriteReportJSON) {
-                            $OverwriteReportJSON = Read-Host -Prompt "A report JSON already exists in the specified folder for $($AsBuiltReportModule.Name). Would you like to overwrite it? (y/n)"
-                        }
-                        if ($OverwriteReportJSON -eq 'y') {
-                            Try {
-                                New-AsBuiltReportConfig -Report $AsBuiltReportName -Path $ReportConfigFolder -Overwrite
-                            } Catch {
-                                Write-Error $_
-                                Break
-                            }
+                if (Test-Path -Path $ReportModulePath) {
+                    $OverwriteReportConfig = Read-Host -Prompt "A report configuration file already exists in the specified folder for $($ReportModule.Name). Would you like to overwrite it? (y/n)"
+                    while ("y", "n" -notcontains $OverwriteReportConfig) {
+                        $OverwriteReportConfig = Read-Host -Prompt "A report configuration file already exists in the specified folder for $($ReportModule.Name). Would you like to overwrite it? (y/n)"
+                    }
+                    if ($OverwriteReportConfig -eq 'y') {
+                        Try {
+                            Write-Verbose -Message "Copying '$($ReportModule.ModuleBase)\$($ReportModule.Name).json' to '$($ReportModulePath)'. Overwriting existing file."
+                            New-AsBuiltReportConfig -Report $Report -FolderPath $ReportConfigFolder -Force
+                        } Catch {
+                            Write-Error $_
+                            Break
                         }
                     }
                 }
@@ -204,33 +201,63 @@ function New-AsBuiltConfig {
     Write-Host '----------------------------------------------' -ForegroundColor Cyan
     Write-Host ' <       As Built Report Configuration      > ' -ForegroundColor Cyan
     Write-Host '----------------------------------------------' -ForegroundColor Cyan
-    $SaveAsBuiltConfig = Read-Host -Prompt "Would you like to save the As Built configuration file? (y/n)"
+    $SaveAsBuiltConfig = Read-Host -Prompt "Would you like to save the As Built Report configuration file? (y/n)"
     while ("y", "n" -notcontains $SaveAsBuiltConfig) {
-        $SaveAsBuiltConfig = Read-Host -Prompt "Would you like to save the As Built configuration file? (y/n)"
+        $SaveAsBuiltConfig = Read-Host -Prompt "Would you like to save the As Built Report configuration file? (y/n)"
     }
 
     if ($SaveAsBuiltConfig -eq 'y') {
-        $AsBuiltName = Read-Host -Prompt "Enter a name for the As Built report configuration file [AsBuiltReport]"
+        $AsBuiltName = Read-Host -Prompt "Enter a name for the As Built Report configuration file [AsBuiltReport]"
         if (($AsBuiltName -like $null) -or ($AsBuiltName -eq "")) {
             $AsBuiltName = "AsBuiltReport"
         }
         if ($Config.UserFolder.Path) {
-            $AsBuiltExportPath = Read-Host -Prompt "Enter the path to save the As Built report configuration file [$($Config.UserFolder.Path)]"
+            $AsBuiltExportPath = Read-Host -Prompt "Enter the path to save the As Built Report configuration file [$($Config.UserFolder.Path)]"
             if (($AsBuiltExportPath -like $null) -or ($AsBuiltExportPath -eq "")) {
                 $AsBuiltExportPath = $Config.UserFolder.Path
             }
+		} elseif ($ReportConfigFilePath) {
+			$ReportConfigFolderPath = Split-Path -Path $ReportConfigFilePath
+			$AsBuiltExportPath = Read-Host -Prompt "Enter the path to save the As Built Report configuration file [$ReportConfigFolderPath]"
+			if (($AsBuiltExportPath -like $null) -or ($AsBuiltExportPath -eq "")) {
+                $AsBuiltExportPath = $ReportConfigFolderPath
+            }
         } else {
-            $AsBuiltExportPath = Read-Host -Prompt "Enter the path to save the As Built report configuration file [$env:USERPROFILE\AsBuiltReport]"
+            $AsBuiltExportPath = Read-Host -Prompt "Enter the path to save the As Built Report configuration file [$env:USERPROFILE\AsBuiltReport]"
             if (($AsBuiltExportPath -like $null) -or ($AsBuiltExportPath -eq "")) {
                 $AsBuiltExportPath = "$env:USERPROFILE\AsBuiltReport"
-                $Config.UserFolder = @{
-                    'Path' = $AsBuiltExportPath
-                }
-            }
+            }          
         }
+		if (!(Test-Path -Path $AsBuiltExportPath)) {
+			Write-Verbose -Message "Creating As Built Report configuration folder '$AsBuiltExportPath'."
+			Try {
+				$Folder = New-Item -Path $AsBuiltExportPath -ItemType Directory -Force
+			} Catch {
+				Write-Error $_
+				break
+			}
+		}
+		$Config.UserFolder = @{
+			'Path' = $AsBuiltExportPath
+		}
+        Write-Verbose -Message "Saving As Built Report configuration file '$($AsBuiltName).json' to path '$AsBuiltExportPath'."
         $AsBuiltConfigPath = Join-Path -Path $AsBuiltExportPath -ChildPath "$AsBuiltName.json"
         $Config | ConvertTo-Json | Out-File $AsBuiltConfigPath
+    } else {
+        Write-Verbose -Message "As Built Report configuration file not saved."
     }
-    $Config
     #endregion Save configuration
+
+    # Print output to screen so that it can be captured to $Global:AsBuiltConfig variable in New-AsBuiltReport
+    $Config
+
+    # Verbose Output
+    Write-Verbose -Message "Config.Report.Author = $ReportAuthor"
+    Write-Verbose -Message "Config.UserFolder.Path = $ReportConfigFolder"
+    foreach ($x in $Config.Company.Keys) {
+        Write-Verbose -Message "Config.Company.$x = $($Config.Company[$x])"
+    }
+    foreach ($x in $Config.Email.Keys) {
+        Write-Verbose -Message "Config.Email.$x = $($Config.Email[$x])"
+    }
 }#End New-AsBuiltConfig Function